requirements.txt
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
fastapi
uvicorn
jinja2
python-multipart
asyncpg
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6